# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: AI Store Manage - 用户与部门管理 API
  description: |
    WMS 仓储管理系统 — 用户与部门模块 API 契约。

    本契约基于 `sys_user` 表字段设计，遵循 Contract First 原则，
    定义用户 CRUD、状态变更、密码重置等接口。

    **设计原则：**
    - RESTful 风格，资源路径统一为 `/api/users`
    - 严格区分数据库字段与 API 暴露字段
    - `password` 仅出现在写入类请求，永不出现在响应中
    - `deleted` 为内部逻辑删除标记，不暴露给业务接口
    - 枚举字段（gender、status）明确约束取值范围
    - 统一错误响应结构，便于前端统一处理
  version: 1.0.0
  contact:
    name: AI Store Manage Team
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
paths:
...
  /api/unit-conversions:
    post:
      tags: [计量换算管理]
      summary: 创建计量换算
      operationId: createUnitConversion
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateUnitConversionRequest' } } }
      responses:
        '201': { description: 创建成功, content: { application/json: { schema: { $ref: '#/components/schemas/UnitConversion' } } } }
        '400': { description: 校验失败, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: SKU 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    get:
      tags: [计量换算管理]
      summary: 查询计量换算列表
      operationId: listUnitConversions
      parameters:
        - { name: skuId, in: query, required: false, schema: { type: integer, format: int64 } }
        - { name: status, in: query, required: false, schema: { type: integer, enum: [0, 1] } }
        - { name: page, in: query, required: false, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { description: 查询成功, content: { application/json: { schema: { $ref: '#/components/schemas/UnitConversionListResponse' } } } }
  /api/unit-conversions/{id}:
    get:
      tags: [计量换算管理]
      summary: 查询计量换算详情
      operationId: getUnitConversionById
      parameters: [ { $ref: '#/components/parameters/UnitConversionId' } ]
      responses:
        '200': { description: 查询成功, content: { application/json: { schema: { $ref: '#/components/schemas/UnitConversion' } } } }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    put:
      tags: [计量换算管理]
      summary: 更新计量换算
      operationId: updateUnitConversion
      parameters: [ { $ref: '#/components/parameters/UnitConversionId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateUnitConversionRequest' } } }
      responses:
        '200': { description: 更新成功, content: { application/json: { schema: { $ref: '#/components/schemas/UnitConversion' } } } }
        '400': { description: 校验失败, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    delete:
      tags: [计量换算管理]
      summary: 删除计量换算
      operationId: deleteUnitConversion
      parameters: [ { $ref: '#/components/parameters/UnitConversionId' } ]
      responses:
        '204': { description: 删除成功 }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/unit-conversions/{id}/status:
    patch:
      tags: [计量换算管理]
      summary: 变更计量换算状态
      operationId: updateUnitConversionStatus
      parameters: [ { $ref: '#/components/parameters/UnitConversionId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateStatusRequest' } } }
      responses:
        '200': { description: 变更成功, content: { application/json: { schema: { $ref: '#/components/schemas/UnitConversion' } } } }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  # ── 物料图片管理 ──
...
components:
  parameters:
...
    UnitConversionId:
      name: id
      in: path
      description: 计量换算 ID
      required: true
      schema: { type: integer, format: int64, example: 1 }
...
  schemas:
...
    UpdateStatusRequest:
      type: object
      description: 通用状态变更请求体（0=禁用,1=启用）
      required: [status]
      properties:
        status: { type: integer, enum: [0, 1], example: 0 }
...
    CreateUnitConversionRequest:
      type: object
      required: [skuId, fromUnit, toUnit, factor]
      properties:
        skuId: { type: integer, format: int64 }
        fromUnit: { type: string, minLength: 1, maxLength: 16, example: "ROLL" }
        toUnit: { type: string, minLength: 1, maxLength: 16, example: "M2" }
        factor: { type: number, description: from→to 换算系数 }
        status: { type: integer, enum: [0, 1], default: 1 }
    UpdateUnitConversionRequest:
      type: object
      properties:
        factor: { type: number }
    UnitConversion:
      type: object
      properties:
        id: { type: integer, format: int64 }
        skuId: { type: integer, format: int64 }
        skuCode: { type: string, nullable: true }
        skuName: { type: string, nullable: true }
        fromUnit: { type: string }
        toUnit: { type: string }
        factor: { type: number }
        status: { type: integer, enum: [0, 1] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        createdBy: { type: integer, format: int64, nullable: true }
        updatedBy: { type: integer, format: int64, nullable: true }
    UnitConversionSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        skuId: { type: integer, format: int64 }
        skuName: { type: string, nullable: true }
        fromUnit: { type: string }
        toUnit: { type: string }
        factor: { type: number }
        status: { type: integer, enum: [0, 1] }
        createdAt: { type: string, format: date-time }
    UnitConversionListResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/UnitConversionSummary' } }
        total: { type: integer, format: int64 }
        page: { type: integer }
        pageSize: { type: integer }
        totalPages: { type: integer }

    # 物料图片（仅存 URL）
...
    ErrorResponse:
      type: object
      description: 统一错误响应体
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 错误码（大写下划线风格，用于前端判断错误类型）
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: 错误描述（人类可读）
          example: "请求参数校验失败"
        details:
          type: array
          description: 字段级错误详情（参数校验失败时返回）
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      description: 字段级错误详情
      properties:
        field:
          type: string
          description: 出错字段名
          example: "username"
        message:
          type: string
          description: 字段错误描述
          example: "用户名长度必须在4-64个字符之间"